Refactor realtime handlers into factories
